functionality completed

+++ src/main/webapp/index.jsp
@@ -10,7 +10,7 @@
        
         FetchCustomers fetchCustomersInstance = new FetchCustomers();
         customers = fetchCustomersInstance.getCustomers();
-      
+   
       }catch(Exception e){
         System.out.println("something went wrong in customers jsp");
         System.out.println(e.getMessage());
@@ -27,7 +27,7 @@
         </button>
         </a>
     <table>
-      <tbody>
+      <thead>
         <tr>
           <th>Customer id</th>
           <th>Customer name</th>
@@ -36,18 +36,30 @@
           <th>Edit</th>
           <th>Delete</th>
         </tr>
+      </thead>
+      <tbody>
+        <%
+        if(customers.isEmpty()){
+         %>  
+        <tr>
+         <td>no customers found</td>
+        </tr>
+       
+        <% } %>
 
         <% for(Customer customer:customers){ %>
 
         <tr>
+
           <td><%= customer.getId() %></td>
           <td><%= customer.getName() %></td>
           <td><%= customer.getAge() %></td>
           <td><%= customer.getRating() %></td>
-        <td><a href="edit?id=<%=customer.getId()"><button>edit</button></a></td>
-        <td><a href="delete?id=<%=customer.getId()"><button>delete</button></a></td>
+        <td><a href="edit?id=<%=customer.getId() %>" ><button>edit</button></a></td>
+        <td><a href="delete?id=<%=customer.getId() %>" onclick="return confirm('are you sure you want to delete <%= customer.getName() %> ? ')"><button>delete</button></a></td>
         </tr>
         <% } %>
+
       </tbody>
     </table>
